+++ Roxane/RXRoxane/RXMove.cpp
@@ -72,12 +72,12 @@ unsigned long long RXMove::random_pick_bit_in_legalmoves(const unsigned long lon
     
     // Masques de zones (à initialiser une seule fois avec les bits correspondants)
     // Chaque bit à 1 représente une case RARE ou MEDIUM
-    static const uint64_t MASK_RARE   = 0x0042000000004200ULL; // B2, G2, B7 & G7
-    static const uint64_t MASK_MEDIUM = 0x4281000000008142ULL; // B1, G1, A2, H2, A7, H7, B8 & G8
+    static const unsigned long long MASK_RARE   = 0x0042000000004200ULL; // B2, G2, B7 & G7
+    static const unsigned long long MASK_MEDIUM = 0x4281000000008142ULL; // B1, G1, A2, H2, A7, H7, B8 & G8
 
     if (legal_moves == 0) return 0;
 
-    std::vector<uint64_t> moves;
+    std::vector<unsigned long long> moves;
     std::vector<int> weights;
     
     // On réserve un peu d'espace pour éviter trop de réallocations
@@ -85,11 +85,11 @@ unsigned long long RXMove::random_pick_bit_in_legalmoves(const unsigned long lon
     weights.reserve(moves.capacity());
 
     // Extraction des bits et attribution des poids
-    uint64_t temp_moves = legal_moves;
+    unsigned long long temp_moves = legal_moves;
     while (temp_moves) {
         // Isoler le bit de poids faible (LSB)
         // Expression mathématique : bit = temp_moves & -temp_moves
-        uint64_t bit = temp_moves & -temp_moves;
+        unsigned long long bit = temp_moves & -temp_moves;
         
         moves.push_back(bit);
